[feat]: restore Direct scenes in seconds hero heading
Replace the center AI Video title with the original animated hero tagline.
